Fixed some critical issue in RX, add support for extended execution
