+++ RX_Explorer/Dialog/QuickStartModifiedDialog.xaml
@@ -162,10 +162,12 @@
                         <Button.Flyout>
                             <Flyout Placement="Top"
                                     x:Name="PickAppFlyout">
-                                <StackPanel>
+                                <StackPanel MaxWidth="230">
                                     <TextBlock Text="您要选择哪一种应用?"
+                                               TextWrapping="WrapWholeWords"
                                                x:Uid="QuickStartPickFileType"/>
                                     <StackPanel Orientation="Horizontal"
+                                                HorizontalAlignment="Center"
                                                 Margin="0,10,0,0">
                                         <Button Content="UWP"
                                                 Width="100"
