+++ RX_Explorer/RX_Explorer.csproj
@@ -167,6 +167,7 @@
     <Compile Include="Class\CustomDataTemplateSelector.cs" />
     <Compile Include="Class\Enum.cs" />
     <Compile Include="Class\Exception.cs" />
+    <Compile Include="Class\ExtendedExecutionController.cs" />
     <Compile Include="Class\FileSystemStorageFile.cs" />
     <Compile Include="Class\FileSystemStorageFolder.cs" />
     <Compile Include="Class\FilterController.cs" />
@@ -184,9 +185,8 @@
     <Compile Include="Class\HiddenStorageFolder.cs" />
     <Compile Include="Class\LinkStorageFile.cs" />
     <Compile Include="Class\InstalledApplication.cs" />
-    <Compile Include="Class\QueueFileOperationController.cs" />
+    <Compile Include="Class\QueueTaskController.cs" />
     <Compile Include="Class\RecycleStorageFolder.cs" />
-    <Compile Include="Class\SystemFontFamilyController.cs" />
     <Compile Include="CustomControl\AppBarButtonWithImage.xaml.cs">
       <DependentUpon>AppBarButtonWithImage.xaml</DependentUpon>
     </Compile>
@@ -751,10 +751,7 @@
       <Version>2.5.0</Version>
     </PackageReference>
     <PackageReference Include="MySqlConnector">
-      <Version>1.3.2</Version>
-    </PackageReference>
-    <PackageReference Include="SharpDX.Direct2D1">
-      <Version>4.2.0</Version>
+      <Version>1.3.3</Version>
     </PackageReference>
     <PackageReference Include="SharpZipLib">
       <Version>1.3.1</Version>
