+++ RX_Explorer/View/TabViewContainer.xaml
@@ -82,13 +82,6 @@
                     </Button>
                 </muxc:TabView.TabStripFooter>
             </muxc:TabView>
-            <controls:InAppNotification StackMode="Replace"
-                                        Name="Notification"
-                                        MinWidth="300"
-                                        x:FieldModifier="public"
-                                        ShowDismissButton="False"
-                                        HorizontalAlignment="Center" />
-
         </Grid>
 
         <SplitView.Pane>
@@ -117,7 +110,7 @@
                 <ListView Grid.Row="1"
                           Margin="0,5,0,0"
                           SelectionMode="None"
-                          ItemsSource="{x:Bind Cls:QueueFileOperationController.ListItemSource}"
+                          ItemsSource="{x:Bind Cls:QueueTaskController.ListItemSource}"
                           ItemContainerTransitions="{x:Bind Cls:AnimationController.Current.RepositionTransitions, Mode=OneWay}">
                     <ListView.ItemContainerStyle>
                         <Style TargetType="ListViewItem">
@@ -215,8 +208,8 @@
                             VerticalAlignment="Bottom">
                     <CheckBox Content="允许并行执行任务"
                               x:Uid="TaskListAllowParalledTask"
-                              IsChecked="{x:Bind Cls:QueueFileOperationController.AllowParalledExecution, Mode=TwoWay}" />
-                    <CheckBox IsChecked="{x:Bind Cls:QueueFileOperationController.OpenPanelWhenTaskIsCreated, Mode=TwoWay}">
+                              IsChecked="{x:Bind Cls:QueueTaskController.AllowParalledExecution, Mode=TwoWay}" />
+                    <CheckBox IsChecked="{x:Bind Cls:QueueTaskController.OpenPanelWhenTaskIsCreated, Mode=TwoWay}">
                         <TextBlock TextWrapping="WrapWholeWords"
                                    VerticalAlignment="Center"
                                    x:Uid="TaskListAlwaysOpenPanel" />
